% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello_world.R
\name{hello_world}
\alias{hello_world}
\title{Hello world}
\usage{
hello_world(who = c("world", "tilt"))
}
\arguments{
\item{who}{A valid string.}
}
\value{
A string.
}
\description{
Hello world
}
\examples{
hello_world()

hello_world("tilt")

# Errors
try(hello_world("invalid"))
}
