% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Say hello}
\usage{
hello(who = c("world", "tilt"))
}
\arguments{
\item{who}{A valid string.}
}
\value{
A string.
}
\description{
Say hello
}
\examples{
hello()

hello("tilt")

# Errors
try(hello("invalid"))
}
