% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\alias{hello_addin}
\title{Greet the world or tilt with code or your mouse}
\usage{
hello(who = c("world", "tilt"))

hello_addin()
}
\arguments{
\item{who}{A valid string.}
}
\value{
A string.
}
\description{
Greet the world from R with \code{hello()} or from the web-browser with
\code{hello_addin()} (from RStudio click on "Addins" and type "hello").
}
\examples{
hello()

hello("tilt")

# Errors
try(hello("invalid"))
}
